+++ src/main/java/com/flavorsujung/isthereopen/respository/BarPatronRepository.java
@@ -4,5 +4,5 @@ import com.flavorsujung.isthereopen.domain.entity.BarPatron;
 import com.flavorsujung.isthereopen.domain.entity.Cafe;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BarPatronRepository extends JpaRepository<BarPatron, Integer> {
+public interface BarPatronRepository extends JpaRepository<BarPatron, Long> {
 }
