+++ src/main/java/com/flavorsujung/isthereopen/respository/CafeRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CafeRepository extends JpaRepository<Cafe,Integer> {
+public interface CafeRepository extends JpaRepository<Cafe,Long> {
    // List<Cafe> findAllCafe();
-    Cafe findCafeBySeq(Integer cafeSeq);
+    Cafe findCafeBySeq(Long cafeSeq);
 }
