+++ src/main/java/com/flavorsujung/isthereopen/respository/RestaurantRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface RestaurantRepository extends JpaRepository<Restaurant,Integer>  {
+public interface RestaurantRepository extends JpaRepository<Restaurant,Long>  {
     //List<Restaurant> findAllRestaurant();
-    Restaurant findRestaurantBySeq(Integer restaurantSeq);
+    Restaurant findRestaurantBySeq(Long restaurantSeq);
 
 }
